Move computation back into compute mod
